Изменены названия файлов view папки, реализован глобальный и локальные фильтры, проделан мелкий рефакторинг
