+++ src/view/profile-level.js
@@ -8,10 +8,12 @@ export const createProfileLevelTemplate = (movies) => {
   }
 
   if (movies > 0) {
-    template = `<section class="header__profile profile">
-    <p class="profile__rating">${rank}</p>
-    <img class="profile__avatar" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
-  </section>`;
+    template = `
+    <section class="header__profile profile">
+      <p class="profile__rating">${rank}</p>
+      <img class="profile__avatar" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
+    </section>
+    `;
   }
 
   return template;
