+++ src/view/stats.js
@@ -1,5 +1,7 @@
 export const createStatsTemplate = (moviesAmount) => {
-  return `<section class="footer__statistics">
-  <p>${moviesAmount} movies inside</p>
-</section>`;
+  return `
+  <section class="footer__statistics">
+    <p>${moviesAmount} movies inside</p>
+  </section>
+  `;
 };
